+++ lib/src/styling/message_style.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:chat_ui_kit/src/widgets/core/messages_list.dart';
 
 enum AvatarBehaviour { alwaysTop, alwaysBottom }
 
+/// Styling and settings for [MessagesList].
 class MessageStyle {
   /// Only used for incoming messages
   final AvatarBehaviour avatarBehaviour;
@@ -16,10 +18,15 @@ class MessageStyle {
 
   /// The foreground color applied to the whole item when it is selected
   /// Defaults to Colors.white.withAlpha(50)
-  final Color selectionColor;
+  final Color? selectionColor;
+
+  /// [ScrollPhysics] parameter for the list of messages.
+  /// Defaults to platform default.
+  final ScrollPhysics? physics;
 
   const MessageStyle(
       {this.padding = EdgeInsets.zero,
+      this.physics,
       this.avatarWidth = 48.0,
       this.selectionColor,
       this.avatarBehaviour = AvatarBehaviour.alwaysBottom});