+++ lib/src/utils/builders.dart
@@ -12,4 +12,4 @@ typedef DateBuilder = Widget Function(BuildContext context, DateTime date);
 /// Builder called to construct parts of the [MessagesListTile] widget.
 /// [index] is the item's position in the list
 typedef MessageWidgetBuilder<T extends MessageBase> = Widget Function(
-    BuildContext context, int index, T item, MessagePosition messagePosition);
+    BuildContext context, int index, T item, MessagePosition? messagePosition);